refactor(Response.kt): 파일 위치 변경
Response.kt 파일 위치 app -> domain 모듈로 이전
